trying to make pins work
