+++ src/components/map/CountriesTileLayer.tsx
@@ -1,10 +1,13 @@
-import L, { FeatureGroup, LeafletEvent, LeafletMouseEvent } from 'leaflet';
+import L, { LeafletEvent, LeafletMouseEvent } from 'leaflet';
 import { useEffect, useState } from "react";
 import ReactDOMServer from 'react-dom/server';
 import { useMap } from "react-leaflet";
 import { LanguageType } from '../../types';
 import { createAltPopup } from "../../utils/mapUtils";
 import { highlightStyles, layerStyle } from './MapStyle';
+// TODO: Write definitions for leaflet vectorgrid ---> Ewan
+// @ts-ignore
+import VectorGrid from "leaflet.vectorgrid";
 
 interface countrySelectionProps {
   url: string
@@ -15,10 +18,11 @@ interface countrySelectionProps {
 
 const toggleCountrySelection = (e: LeafletEvent, layer: any, highlightStyle: any) => {
   // only change selection style on countries that have seroprev data
-  const layerStyle = layer._overriddenStyles[e.layer.properties.CODE]; //TODO: refactor this kinda ugly
+  const layerStyle = layer._overriddenStyles[e.propagatedFrom.properties.CODE]; //TODO: refactor this kinda ugly
 
-  if (layerStyle)
+  if (layerStyle) {
     layer.setFeatureStyle(e.layer.properties.CODE, highlightStyle);
+  }
 }
 
 export default function CountriesTileLayer(props: countrySelectionProps) {
@@ -26,10 +30,11 @@ export default function CountriesTileLayer(props: countrySelectionProps) {
   const { url, records, zIndex, language } = props;
   const map = useMap();
   const [layer, setLayer] = useState(undefined as any);
-  const Lcopy = L as any
   useEffect(() => {
-    var layer = Lcopy.vectorGrid.protobuf(`${url}/tile/{z}/{y}/{x}.pbf`, {
-      rendererFactory: Lcopy.svg.tile,
+    // @ts-ignore
+    const layer = L.vectorGrid.protobuf(`${url}/tile/{z}/{y}/{x}.pbf`, {          
+      // @ts-ignore
+      rendererFactory: L.svg.tile,
       attribution: '',
       interactive: true,
       vectorTileLayerStyles: layerStyle,
@@ -41,7 +46,7 @@ export default function CountriesTileLayer(props: countrySelectionProps) {
     ).addTo(map);
 
     setLayer(layer);
-  }, [Lcopy.svg.tile, Lcopy.vectorGrid, map, url, zIndex])
+  }, [map, url, zIndex])
 
   useEffect(() => {
     if (layer) {
@@ -68,23 +73,24 @@ export default function CountriesTileLayer(props: countrySelectionProps) {
             layer.openPopup(e.latlng);
             const country = records.find(x => x.alpha3Code === e.sourceTarget.properties.CODE)
             if (country) {
-              var test2 = ReactDOMServer.renderToString(createAltPopup(country, language))
-              pop.setContent(test2);
+              pop.setContent(ReactDOMServer.renderToString(createAltPopup(country, language)));
             }
           }
 
           toggleCountrySelection(e, e.target, highlightStyles.hovering)
         },
         mouseout: (e: LeafletMouseEvent) => {
-          if (layer.getPopup())
+          if (layer.getPopup()) {
             layer.closePopup();
+          }
 
           toggleCountrySelection(e, e.target, highlightStyles.default)
         },
         mousemove: (e: LeafletMouseEvent) => {
           const pop = layer.getPopup()
-          if (pop && e.latlng)
+          if (pop && e.latlng) {
             pop.setLatLng(e.latlng)
+          }
         },
         click: () => { }
       })
